Scripts comentados, queda limpiar el index.js

+++ Web-app/Analysis-app/general/escoba.sh
@@ -2,7 +2,8 @@
 
 cd ./Analysis-app/
 
-rm -f general/global.json
+# Limpiar archivos de resultados de ejecuciones anteriores
+if [[ $# -eq 0 ]]; then rm -f general/global.json; fi
 rm -f SSH/enumedSSH.json
 rm -f SAMBA/enumedSamba.json
 rm -f FTP/enumedFTP.json
